requirements.txt
openai
aiofiles
Pillow
python-dotenv
fastapi
uvicorn
pydantic